Consolidate STG105 tests into shared per-client test classes
Moves getLayout/data-locality tests out of standalone files (which
deviated from this repo's established convention of one shared
per-client test class holding all operation tests) into the
conventional homes:

- BlobClientBaseGetLayoutApiTests -> BlobApiTests
- BlobClientBaseGetLayoutAsyncApiTests -> BlobAsyncApiTests
- BlobDataLocalityDownloadApiTests -> BlobApiTests (sync only; no
  async openInputStream/data-locality counterpart exists to move)
- FileGetLayoutTests -> FileApiTest
- FileGetLayoutAsyncTests -> FileAsyncApiTests
- FileDataLocalityDownloadTests -> FileApiTest (sync only, same reason)

Test bodies were also adapted to reuse each target class's existing
conventions instead of introducing new ones: shared @BeforeEach-uploaded
default content where sufficient, existing getRandomFile/compareFiles
helpers for larger multi-chunk scenarios, and existing
modifiedMatchAndLeaseIdSupplier/invalidModifiedMatchAndLeaseIdSupplier
access-condition suppliers already defined in FileApiTest/
FileAsyncApiTests instead of duplicating them.

Recording filenames are keyed by class name, so this move requires a
fresh RECORD pass -- deferred to a follow-up commit once the remaining
STG105 test-coverage gaps (see next commits) are also closed, so
recording only needs to happen once.

Copilot-Session: c4f5bab9-3ca0-444d-97cb-328fb210e7d7
